Stream Y: 31 P1 items — grad-flow, parity, precision, GIL, unchecked async
audit-5 P1.  Lands every Stream Y item (Y.11 + Y.16 already with X.7).

Autograd (Y.1–Y.9):
* Y.1 jvp_norm helper sig float → double; adapter no longer narrows
  the user's AttrKey::P double through the |x|^(p-1) / y^(1-p)
  chain.  The unavoidable narrow at tenzor::norm primal is the
  public-API holdout (out of Y.1 scope).
* Y.2 jvp_index_fill / jvp_masked_fill float value → double;
  adapters pass attrs.get_float(...) directly.
* Y.3 public masked_fill / index_fill gain `double` overloads;
  `float` overloads kept as thin forwarders with @deprecated.
* Y.4 WhereBackward saves x_shape / y_shape on forward;
  backward + backward_with_variables route grads through
  reduce_grad_for_broadcasting / reduce_grad_var_for_broadcasting
  matching BinaryOp backwards.  Mirror-broadcast where(cond[B,N],
  x[1,N], y[B,1]) now produces correct shape-preserving grads.
* Y.5 autograd::chunk emits `chunks` zero-sized slice(input, dim,
  0, 0) outputs when dim_size == 0 (PyTorch parity).  Fixes
  destructured-binding OOB on `auto [a,b,c] = chunk(zeros({0,5}),
  3)`.
* Y.6 select_scatter / slice_scatter normalise dim += ndim with
  range-check at the dispatcher before setting AttrKey::Dim.
* Y.7 same for kthvalue / cummin / cummax dispatchers.
* Y.8 src/ops/async_ops.cpp StreamManager: cudaSetDevice /
  cudaStreamCreate / cudaStreamSynchronize wrapped in a local
  TENZOR_ASYNC_CUDA_CHECK macro (mirrors tenzor::cuda::TENZOR_
  CUDA_CHECK without kernel-side intrinsics).  Destructor logs
  via fprintf since it can't throw.
* Y.9 verified: Mul/Div/MatMul Wirtinger paths already use
  saved_variables_[i] directly when has_saved_variables() is
  true (X.5 + Y.9 in place from prior commits); no regression.

Backend kernels (Y.10–Y.16):
* Y.10 new vulkan::ensure_atomic_int64_supported(device_id,
  op_name) + VulkanBackend::has_atomic_int64() accessor (mirrors
  R.13 / S.4 / U.5 / U.6 / U.7).  Gated at MaxPool2dBackward x2,
  MaxPool1dBackward, AdaptiveMaxPool1dBackward, FractionalMaxPool
  2d/3dBackward, FFT Put.  ROIAlignBackward gated on
  ensure_atomic_float_supported (its f64 shader uses
  GL_EXT_shader_atomic_float, not _int64).  Non-atomic f64
  shaders (max_pool2d_backward_indices_f64, max_pool3d_backward_
  f64) intentionally skipped.
* Y.11 + Y.16 landed with X.7 (Vulkan F16/BF16 CAS sweep across
  43 shaders).
* Y.12 OneAPI fold_kernel / unfold_kernel widened scalar →
  per-axis (kH, kW, sH, sW, pH, pW, dH, dW); col2im/im2col index
  math rewritten per-axis.  Drops asymmetric-rejection throws.
  OneAPI now leads on Fold/Unfold parity.
* Y.13 ROCm split_kernel normalises negative dim with range check;
  drops the trailing hipStreamSynchronize that defeated async
  pipelining.
* Y.14 every queue.wait() → queue.wait_and_throw() across
  src/backends/oneapi/{kernels/{sycl_sort_utils.hpp, indexing,
  grid_sample, nested_attention, stft, sampling, reduction,
  conv2d, fused_ops, statistical, batchnorm, transform, math},
  oneapi_kernel_registry}.cpp.  Async OOM / oversubscription /
  kernel timeout no longer silently dropped.
* Y.15 ROCm fp16_saturate_kernel checks raw F16 bit pattern
  (exp==0x1F && mant != 0) BEFORE __half2float; on NaN, skip the
  clamp so NaN payload bits remain untouched.  Mirrors
  rocm_nan_helpers.hip.h::is_nan_bits.

nn / optim / distributed (Y.17–Y.22):
* Y.17 NAdam / LAMB load_state_dict adopt the V.27 cast-on-load
  pattern via the shared master_weights.hpp helper.  Lion already
  had it.
* Y.18 Adam (+ AdamW) load_state_dict pre-allocates
  max_exp_avg_sq_ via make_optim_state when the checkpoint's
  amsgrad=true flips from a constructor-default amsgrad=false.
  No more silent discard.
* Y.19 R.21 attn_mask Bool / integer dtype now normalised to
  Float32 + where(mask, -inf, 0) at the top of the reshape branch
  (mirrors V.28 key_padding_mask).  All 3 ranks (2D/3D/4D)
  correctly produce -inf at masked positions.
* Y.20 InstanceNorm1d / InstanceNorm2d / GroupNorm forward gain
  results.size() < 3 throw before reading results[1]/results[2]
  (mirrors V.24 LayerNorm).
* Y.21 Module::load_state_dict submodule recursion now passes
  the root's missing_keys / unexpected_keys into the W.17
  out-param overload, re-prefixed with `name + "."`.  Strict mode
  now sees submodule mismatches.
* Y.22 FSDP2 shard_parameters() pre-allocates unsharded_dst_
  with zeros(full_shape, slot_dtype, device) for every param
  when shard_world_size_ > 1.  Cycle 0 hits the in-place
  zero_+add_ path instead of the TensorImpl swap.

Python frontend + GradScaler (Y.23–Y.27, Y.32):
* Y.23 Variable __getitem__ ports V.35 TupleKinds (NewAxis,
  Ellipsis, BoolMask).  Bare var[None] / var[...] synthesize
  one-entry tuple.  Execution routes through
  tenzor::unsqueeze(Variable), tenzor::index_select(Variable),
  tenzor::masked_select(Variable, Variable), tenzor::slice(
  Variable), tenzor::squeeze(Variable) — grad_fn preserved.
* Y.24 functional_{batch,layer,group,instance,rms}_norm gain
  optional weight + bias Variable kwargs.  C++ bindings apply
  affine via Variable mul + add (channel-first reshape [C] →
  [1,C,1,…] for BN/GN/IN; trailing broadcast for LN/RMS) so
  grads flow.  Python wrappers in functional.py thread the
  kwargs through.
* Y.25 create_numpy_array signature gains want_no_copy=false;
  strided-fallback path throws ValueError when want_no_copy is
  true instead of the W.16-era warn+copy.
* Y.26 jit.trace / CompiledModule.forward / __call__ /
  trace_module gain py::call_guard<gil_scoped_release>().
  jit.compile's returned cpp_function already released inside.
* Y.27 already closed in X.9 (autograd.pyi declares FunctionCtx;
  no_grad/enable_grad re-exported with __all__).
* Y.32 GradScaler gains private side-table f32_unscaled_grads_;
  unscale_ stores F32 unscaled tensor before casting back;
  check_inf_nan_ consults side-table first (falls back to
  *grad for native-F32 params); step() / reset() clear the
  table.

Tests / build (Y.28–Y.31, Y.33, Y.34):
* Y.28 parity_coverage_matrix.py exits non-zero with explicit
  "run from build/ directory" message when ctest --show-only=json
  -v1 returns zero tests.  No more silent empty-matrix mimicry.
* Y.29 CMakeLists.txt: TENZOR_TEST_BFLOAT16 default flipped to
  ON for non-Debug builds; Debug keeps OFF for dev iteration
  speed.  Documented in TESTING.md.
* Y.30 tests/backend_parity/test_nn_conv_parity.cpp gains
  Conv1d_PerAxisStride / Conv1d_PerAxisDilation TEST_Ps using
  the std::vector<int64_t> overload.
* Y.31 test_nn_loss_parity.cpp CTCLoss test no longer wraps in
  try/...GTEST_SKIP — real failures surface.
* Y.33 tests/utils/test_{logging, tensorboard, config}.cpp +
  tests/unit/test_callbacks.cpp /tmp paths converted to
  temp_directory_path() / "tenzor_<name>_<pid>_<test>" mirroring
  T.11 ONNXImportTest fix.
* Y.34 tests/CMakeLists.txt:25 `off` → `OFF` for consistency.

Verified: ninja build clean (single -j1 shell).  Tooling gates
verified after each fix:
  python3 tools/scan_audit_markers.py         → 0
  python3 tools/check_pyi_drift.py            → 0
  python3 tools/check_multidtype_coverage.py  → 0
  python3 tools/audit_tolerances.py --strict  → 0
  python3 tools/check_param_group_serial....py → 0
  python3 tools/check_tensorboard_wire_form.py → 0
  python3 tools/parity_coverage_matrix.py     → 0
  python3 tools/lint_test_contract.py         → 0

+++ include/tenzor/autograd/jvp_rules.hpp
@@ -177,7 +177,7 @@ auto jvp_nan_to_num(const DualTensor& x, double nan, double posinf, double negin
 /// @name Reductions long-tail (Audit A.4 batch 5)
 /// @{
 /// p-norm: tangent = sum(sign(x) * |x|^(p-1) * dx, dim) * norm^(1-p)
-auto jvp_norm(const DualTensor& x, float p, std::optional<int64_t> dim, bool keepdim) -> DualTensor;
+auto jvp_norm(const DualTensor& x, double p, std::optional<int64_t> dim, bool keepdim) -> DualTensor;
 /// argmax/argmin/argsort produce integer indices: tangent is zero (not differentiable).
 auto jvp_argmax(const DualTensor& x, std::optional<int64_t> dim, bool keepdim) -> DualTensor;
 auto jvp_argmin(const DualTensor& x, std::optional<int64_t> dim, bool keepdim) -> DualTensor;
@@ -197,7 +197,7 @@ auto jvp_index_copy(const DualTensor& input, int64_t dim, const Tensor& index,
                     const DualTensor& source) -> DualTensor;
 /// index_fill: y[index[i]] = constant; linear in input with the indexed slots zeroed.
 auto jvp_index_fill(const DualTensor& input, int64_t dim, const Tensor& index,
-                    float value) -> DualTensor;
+                    double value) -> DualTensor;
 /// select_scatter: linear in input and src.
 auto jvp_select_scatter(const DualTensor& input, const DualTensor& src,
                         int64_t dim, int64_t index) -> DualTensor;
